backend/plugins/gemini-tagger/requirements.txt

google-genai
pydantic